Issues Resolved
Resolved a few issues where the Spotify Login was having an error authenticating and forced the app to move to the next activity. Also commented out unneeded intent filters in the Manifest to prevent unneeded icons appearing on the device running the app. Also fixed a few names in the files that needed to match up to run with other files correctly.